:truck: renames utils to lib
