+++ lib/range.ts
